fix: allow 'self' scripts for csp
Refs: OPS-10268

Not sure if this nginx config is being used in production but it appears
to be. The csp config we're using elsewhere is listed at https://docs.google.com/spreadsheets/d/1aJtXMCckhxRSLaMToBS37nwd9XLWeGdtD1rOAbXI7jc/edit#gid=0
With the other sites, we're only reporting csp exceptions, not actually
blocking anything.